refactor(node/sync): extract blob_fetch submodule from the manager god-file (increment 2)
Continues the sync/manager split. Moves the blob / application fetching
cluster into `blob_fetch.rs` — behavior-preserving, pure code movement
(mod.rs: +1 wiring line, -183).

Moved (an `impl SyncManager` fragment):
- get_blob_info, get_application_size, get_application_source,
  install_bundle_after_blob_sharing

Wiring: methods called from the DAG-sync flow in mod.rs are `pub(super)`;
get_application_source stays private (used only within the cluster). Methods
that stay in mod.rs remain reachable via ancestor privacy. Explicit imports,
no `use super::*`.

Validated: node lib tests (333) and sync_sim (314) pass.
